+++ R/connected.R
@@ -5,7 +5,7 @@
 #' @param n Number of lines to start. Default 100.
 #' @param n_geom Number of points along path to create. Default 2.
 #' @param random Boolean value for whether to randomise plotting order of edges.
-#' @param col_palette Colour palette from RColorBrewer. Default "RdPu".
+#' @param col_palette Vector of colours. Default "RdPu" colour palette from RColorBrewer.
 #' @param bg_col Background colour. Default "#ae217e".
 #' @param s Seed value. Default 1234.
 #' @return A ggplot object.
@@ -15,7 +15,7 @@
 connected <- function(n = 100,
                       n_geom = 2,
                       random = FALSE,
-                      col_palette = "RdPu",
+                      col_palette = RColorBrewer::brewer.pal(n = 9, "RdPu"),
                       bg_col = "#ae217e",
                       s = 1234) {
   n <- round(n)
@@ -42,7 +42,7 @@ connected <- function(n = 100,
   p <- ggraph::ggraph(g2, layout = "linear", circular = TRUE) +
     ggraph::geom_edge_arc(ggplot2::aes(colour = .data$e_val), n = n_geom) +
     ggplot2::coord_fixed() +
-    ggraph::scale_edge_colour_distiller(palette = col_palette) +
+    ggraph::scale_edge_colour_gradientn(colours = col_palette) +
     ggplot2::theme(panel.background = ggplot2::element_rect(fill = bg_col, colour = bg_col),
           plot.background = ggplot2::element_rect(fill = bg_col, colour = bg_col),
           plot.title = ggplot2::element_blank(),