+++ NEWS.md
@@ -1,3 +1,10 @@
+## aRt 1.2.0
+
+* add `streams()` function
+* Move {MetBrewer} to Suggests
+* Move {RColorBrewer} to Suggests
+* Functions take colour vectors instead of palette names as inputs
+
 ## aRt 1.1.9
 
 * update license